refactor: Command 클래스를 controller 패키지로 이동
